refactor(model): restructure Coffee.kt and add Review.kt to align with Lei model

+++ app/src/main/java/com/android/brewr/model/coffee/Review.kt
@@ -1,3 +1,3 @@
 package com.android.brewr.model.coffee
 
-data class Review(val authorName: String, val review: String, val rating: Float)
+data class Review(val authorName: String, val review: String, val rating: Double)
